backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
bcrypt==4.1.2
pydantic[email]>=2.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator>=2.0.0
PyJWT==2.8.0
motor==3.3.2
pandas==2.1.4
openpyxl==3.1.2
cryptography==41.0.8
httpx==0.25.2
PyPDF2==3.0.1 